Api has been created for this app using JWT passport

+++ controllers/api/v1/post_api.js
@@ -18,11 +18,19 @@ module.exports.index= async function(req,res){
 module.exports.destroy= async function(req, res) {
     try{
     let post =await Post.findById(req.params.id);
-            post.remove();
-            await Comment.deleteMany({ post: req.params.id });
-            return res.json(200,{
-                message:"Post Deleted"
-            });
+    if(post.user == req.user.id){
+        post.remove();
+        await Comment.deleteMany({ post: req.params.id });
+        return res.json(200,{
+            message:"Post Deleted"
+        });
+    }
+    else{
+        return res.json(401,{
+            message:"You cannot deleate this post"
+        });
+    }
+            
     }catch(err){
         console.log("#############Error",err);
         return res.json(500,{
